feat(deskotp): remember discover language
