+++ apps/desktop/layer/renderer/src/modules/trending/index.tsx
@@ -6,7 +6,7 @@ import { useQuery } from "@tanstack/react-query"
 import { useState } from "react"
 import { useTranslation } from "react-i18next"
 
-import { useGeneralSettingKey } from "~/atoms/settings/general"
+import { setUISetting, useUISettingKey } from "~/atoms/settings/ui"
 import { apiClient } from "~/lib/api-fetch"
 
 import { TrendingFeedCard } from "../discover/TrendingFeedCard"
@@ -52,17 +52,16 @@ export function Trending({
 }) {
   const { t } = useTranslation()
   const { t: tCommon } = useTranslation("common")
-  const lang = useGeneralSettingKey("language")
+  const lang = useUISettingKey("discoverLanguage")
 
-  const [selectedLang, setSelectedLang] = useState<Language>(lang.startsWith("zh") ? "all" : "eng")
   const [selectedView, setSelectedView] = useState<View>("all")
 
   const { data, isLoading } = useQuery({
-    queryKey: ["trending", selectedLang, selectedView],
+    queryKey: ["trending", lang, selectedView],
     queryFn: async () => {
       return await apiClient.trending.feeds.$get({
         query: {
-          language: selectedLang === "all" ? undefined : selectedLang,
+          language: lang === "all" ? undefined : lang,
           view: selectedView === "all" ? undefined : Number(selectedView),
           limit,
         },
@@ -86,9 +85,9 @@ export function Trending({
           <div className="flex items-center">
             <span className="text-text shrink-0 text-sm font-medium">{t("words.language")}:</span>
             <ResponsiveSelect
-              value={selectedLang}
+              value={lang}
               onValueChange={(value) => {
-                setSelectedLang(value as Language)
+                setUISetting("discoverLanguage", value as Language)
               }}
               triggerClassName="h-8 rounded border-0"
               size="sm"
